app/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

scikit-learn==1.1.0
flask-sqlalchemy==2.5.1
psycopg2==2.9.3
flask-login==0.6.2
python-dotenv==0.19.2

# app/app.py: 1
Flask == 2.0.2

# app/helpers/test_helpers.py: 2
Werkzeug == 2.0.2

# app/app.py: 4
joblib == 1.2.0

# app/app.py: 3
pandas == 1.4.0
